Only set the gig if it actually exists.
Will need to add some consideration for 404s later. The reason this is required is that polymer fire will return an empty object if you point `firebase-document` at a non-existent path. Further changes to that object will be attempted to be written to that path. This is handy if you want automatic binding but it’s troublesome if you need to know if something was there to begin with.